refactor: extract store glossary
